Move custom assertions ShouldSatisfy and ShouldBeGenericType to the root of the Fixie.Tests project, so they are no longer colocated with methods deprecated by Fixie.Assertions.
